+++ CHANGELOG.md
@@ -6,7 +6,9 @@ All notable changes to this project will be documented in this file. This projec
 
 [Full Changelog](https://github.com/Shatur95/optimus-manager-qt/compare/1.1.2...HEAD)
 
-_No unreleased changes yet_
+**Changed**
+
+-   Remove "The application is already running" message.
 
 ## [1.1.2](https://github.com/Shatur95/optimus-manager-qt/tree/HEAD)) (2019-04-20)